refactor: code style & help information show &show style
