+++ cmd/info_cmd.go
@@ -20,17 +20,16 @@ package cmd
 
 import (
 	"bytes"
-	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"io/ioutil"
+	"net/http"
 	"os"
+	"strconv"
 
-	sdkutils "github.com/ontio/ontology-go-sdk/utils"
 	"github.com/ontio/ontology/cmd/utils"
-	"github.com/ontio/ontology/common"
 	"github.com/ontio/ontology/common/log"
 	"github.com/urfave/cli"
-	"errors"
 )
 
 var (
@@ -48,31 +47,6 @@ var (
 		Description: ``,
 	}
 )
-
-func showInfoHelp() {
-	var infoHelp = `
-   Name:
-      ontology info                    Show blockchain information
-
-   Usage:
-      ontology info [command options] [args]
-
-   Description:
-      With ontology info, you can look up blocks, transactions, etc.
-
-   Command:
-      version
-
-      block
-         --hash value                  block hash value
-         --height value                block height value
-
-      tx
-         --hash value                  transaction hash value
-
-`
-	fmt.Println(infoHelp)
-}
 
 func infoCommand(context *cli.Context) error {
 	showInfoHelp()
@@ -145,21 +119,6 @@ var versionCommand = cli.Command{
 	Description:  ``,
 }
 
-func showVersionInfoHelp() {
-	var versionInfoHelp = `
-   Name:
-      ontology info version            Show ontology node version
-
-   Usage:
-      ontology info version
-
-   Description:
-      With this command, you can look up the ontology node version.
-
-`
-	fmt.Println(versionInfoHelp)
-}
-
 func versionInfoCommand(ctx *cli.Context) error {
 	version, err := ontSdk.Rpc.GetVersion()
 	if nil != err {
@@ -170,111 +129,43 @@ func versionInfoCommand(ctx *cli.Context) error {
 	return nil
 }
 
-func showBlockInfoHelp() {
-	var blockInfoHelp = `
-   Name:
-      ontology info block             Show blockchain information
-
-   Usage:
-      ontology info block [command options] [args]
-
-   Description:
-      With this command, you can look up block information.
-
-   Options:
-      --hash value                    block hash value
-      --height value                  block height value
-`
-	fmt.Println(blockInfoHelp)
+func txInfoCommand(ctx *cli.Context) error {
+	if ctx.IsSet(utils.HashInfoFlag.Name) {
+		txHash := ctx.String(utils.HashInfoFlag.Name)
+		resp, err := request("GET", nil, restfulAddr()+"/api/v1/transaction/"+txHash)
+		if err != nil {
+			return err
+		}
+		echoBlockGracefully(resp)
+		return nil
+	}
+	showTxInfoHelp()
+	return nil
 }
 
 func blockInfoCommand(ctx *cli.Context) error {
 	if ctx.IsSet(utils.HeightInfoFlag.Name) {
 		height := ctx.Int(utils.HeightInfoFlag.Name)
-		fmt.Println("blockInfo height: ", height)
 		if height >= 0 {
-			block, err := ontSdk.Rpc.GetBlockByHeight(uint32(height))
-
+			resp, err := request("GET", nil, restfulAddr()+"/api/v1/block/details/height/"+strconv.Itoa(height))
 			if err != nil {
-				log.Fatalf("Get block by height(%d) is error:%s", height, err.Error())
 				return err
 			}
-			if block == nil || block.Header == nil {
-				log.Fatalf("Get block by height(%d), the block or block.Header is nil", height)
-				return errors.New("GetBlockByHeight: the block or block.Header is nil ")
-			}
-
-			echoBlockGracefully(block)
+			echoBlockGracefully(resp)
 			return nil
 		}
 	} else if ctx.IsSet(utils.HashInfoFlag.Name) {
 		blockHash := ctx.String(utils.HashInfoFlag.Name)
-		fmt.Println("blockInfo blockHash: ", blockHash)
 		if "" != blockHash {
-			var hash common.Uint256
-			hex, err := hex.DecodeString(blockHash)
+			resp, err := request("GET", nil, restfulAddr()+"/api/v1/block/details/hash/"+blockHash)
 			if err != nil {
-				log.Fatalf("Decode string error, blockHash:%s, err:%s", blockHash, err.Error())
 				return err
 			}
-			if err := hash.Deserialize(bytes.NewReader(hex)); err != nil {
-				log.Fatalf("Deserialize hex error,hex:%s, err:%s", hex, err.Error())
-				return err
-			}
-			block, err := ontSdk.Rpc.GetBlockByHash(hash)
-			if err != nil {
-				log.Fatalf("GetBlock GetBlockFromStore BlockHash:%x error:%s", hash, err)
-				return err
-			}
-			if block == nil || block.Header == nil {
-				return errors.New("GetBlockByHash: the block or block.Header is nil ")
-			}
-
-			echoBlockGracefully(block)
+			echoBlockGracefully(resp)
 			return nil
 		}
 	}
 	showBlockInfoHelp()
-	return nil
-}
-
-func showTxInfoHelp() {
-	var txInfoHelp = `
-   Name:
-      ontology info tx               Show transaction information
-
-   Usage:
-      ontology info tx [command options] [args]
-
-   Description:
-      With this command, you can look up transaction information.
-
-   Options:
-      --hash value                   transaction hash value
-
-`
-	fmt.Println(txInfoHelp)
-}
-
-func txInfoCommand(ctx *cli.Context) error {
-	if ctx.IsSet(utils.HashInfoFlag.Name) {
-		txHash := ctx.String(utils.HashInfoFlag.Name)
-		ontInitTx, err := sdkutils.ParseUint256FromHexString(txHash)
-		if err != nil {
-			log.Errorf("ParseUint256FromHexString error:%s", err)
-			return err
-		}
-
-		tx, err := ontSdk.Rpc.GetRawTransaction(ontInitTx)
-		if err != nil {
-			log.Errorf("GetRawTransaction error:%s", err)
-			return err
-		}
-
-		echoBlockGracefully(tx)
-		return nil
-	}
-	showTxInfoHelp()
 	return nil
 }
 
@@ -290,4 +181,45 @@ func echoBlockGracefully(block interface{}) {
 		log.Fatalf("Gracefully format json err: %s", err.Error())
 	}
 	out.WriteTo(os.Stdout)
+}
+
+func request(method string, cmd map[string]interface{}, url string) (map[string]interface{}, error) {
+	hClient := &http.Client{}
+	var repMsg = make(map[string]interface{})
+	var response *http.Response
+	var err error
+	switch method {
+	case "GET":
+		req, err := http.NewRequest("GET", url, nil)
+		if err != nil {
+			return repMsg, err
+		}
+		response, err = hClient.Do(req)
+	case "POST":
+		data, err := json.Marshal(cmd)
+		if err != nil {
+			return repMsg, err
+		}
+		reqData := bytes.NewBuffer(data)
+		req, err := http.NewRequest("POST", url, reqData)
+		if err != nil {
+			return repMsg, err
+		}
+		req.Header.Set("Content-type", "application/json")
+		response, err = hClient.Do(req)
+	default:
+		return repMsg, err
+	}
+	if response != nil {
+		defer response.Body.Close()
+
+		body, _ := ioutil.ReadAll(response.Body)
+		if err := json.Unmarshal(body, &repMsg); err == nil {
+			return repMsg, err
+		}
+	}
+	if err != nil {
+		return repMsg, err
+	}
+	return repMsg, err
 }
